requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
dnspython==2.4.2
python-whois==0.8.0